Pseudocoding for robot elevator floor input
Planned an algorithms for a robot who receives an user input floor
number and the output result. The robot will need to go through a method
to perform the task. It will run a loop within the method after
receiving an input until it equals the floor number the user inputted.