Add a loop for allergies question
The allergies question will loop until they type done. Each of the
answers will be put into one array that will store all of its string.
It will also delete the string 'done' from the array when it is finish.
If the user types sunshine, it will end the loop, go to the end
of the line and print "probably a vampire."