Add encryption method to feature
The program will ask the user for their pw. Then it will pass the pw
into the encrypt_password method. The method will deconstruct the pw.
Using the for loop base on the length of the pw, it will identify
each letter based on the pw index, move a letter forward, then store it
to an empty variable.