Add a decoder method
The decoder will take the encrypted message, measure the length of
the pw, reverse the index order starting at -1 to scan the last letter
of the pw. Afterward, match the index numbers to the alphabet
string to identify the letter preceding the original letter. Then
stored each letter into a variable call decoded_password with the
variable string reverse to display the correct value.