Add a method that move name letters to next value
The method move_letter takes a name as an argument, splits the letters
and make it into an array. Then it is broken down into individual
string that is passed into a case function, it would then move
each letter based on where it is matched and return the modified name.