Add method to store questions and continue option
The method user_input house the questions for first name and last name.
After receiving the values, it will run it into the move_letter method
and return a new value. Afterward, the method will ask if you want to
continue or not. The loop will go on until it is quit.