Add data structure to the end of user_input method
The hash data structure is going to take the key and value from the
user inputs and store them. It will then print out the user modified
name with user real name.
A fix was made to original_full_name and new_full_name to make the
string output clean.