Add a method to find number in array
The method search_array takes in an array and a search number. It then
look at the length of the array minus 1. This is used to make another
array to store the fake index. The two array will be combine into
a hash with the argument array as the key and the fake index array as
values. The search value will look into the key and pull out the
index that correspond to the key.