Add random santa generator
The program will ask the user for how many santa they want to
generate. It will take the user input and then loop it through
that amount of time. Each time generating a random santa gender
and ethnicity that will be put into random_santas data base. The
program will end with printing the attributes of the generated santa
age, gender, ethnicity.